requirements.txt
pandas
json
streamlit
plotly
python-dotenv
requests


